move global conftest
