+++ protostar/cli/argument_parser_facade_test.py
@@ -3,16 +3,15 @@ from typing import Any, List, Optional, Pattern
 
 import pytest
 
-from conftest import BaseTestCommand, FooCommand
-from protostar.cli.argument import Argument
-from protostar.cli.argument_parser_facade import (
+from .argument import Argument
+from .argument_parser_facade import (
     ArgumentParserFacade,
     MissingRequiredArgumentException,
 )
-from protostar.cli.cli_app import CLIApp
-from protostar.cli.command import Command
-
+from .cli_app import CLIApp
+from .command import Command
 from .config_file_argument_resolver import ConfigFileArgumentResolverProtocol
+from .conftest import BaseTestCommand, FooCommand
 
 
 def test_bool_argument_parsing(foo_command: FooCommand):
